backend/requirements.txt
fastapi>=0.115
uvicorn>=0.30
python-jose[cryptography]>=3.3
pydantic>=2.7
pytest>=8.0
redis>=5.0
